preferences: 20-debian-sid: deprefer t64-specific packages
Current pin set-up allows installing packages from sid only when
they're new (so there is no match in testing). This is what we actually
want (until we get to do the final move to sid) but it also means
that we get into the crosshairs of the t64 migration (as new packages
are being pulled in).

Deprefer *t64 packages, so that they will only be installed if available
in testing.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>